+++ Changelog.md
@@ -5,7 +5,7 @@
 Breaking Changes
 
 *  `PrimitiveContent`’s methods take `self` instead of `&self`. This
-   avoids the lifetime argument in `Primitive`, its encoder. [(#6)]
+   avoids the lifetime argument in `Primitive`, its encoder. [(#7)]
 
 New
 
@@ -13,7 +13,7 @@ Bug Fixes
 
 Dependencies
 
-[(#6)]: https://github.com/NLnetLabs/bcder/pull/6
+[(#7)]: https://github.com/NLnetLabs/bcder/pull/7
 
 
 ## 0.1.0