+++ Changelog.md
@@ -13,11 +13,12 @@ Bug Fixes
 Other Changes
 
 * The `xerr!` macro now prints a backtrace with the `extra-debug` feature
-  enabled instead of panicking. [(#XX)]
+  enabled instead of panicking. [(#21)]
 
 Dependencies
 
 [(#20)]: https://github.com/NLnetLabs/rpki-rs/pull/20
+[(#21)]: https://github.com/NLnetLabs/rpki-rs/pull/21
 
 
 ## 0.2.1