+++ Changelog.md
@@ -4,6 +4,8 @@
 
 Breaking
 
+* The minimum supported Rust version is now 1.34.0. [(#22)]
+
 New
 
 * New methods for checking the class and number on `Tag`. [(#20)]
@@ -19,8 +21,10 @@ Other Changes
 
 Dependencies
 
+
 [(#20)]: https://github.com/NLnetLabs/rpki-rs/pull/20
 [(#21)]: https://github.com/NLnetLabs/rpki-rs/pull/21
+[(#22)]: https://github.com/NLnetLabs/rpki-rs/pull/22
 [(#23)]: https://github.com/NLnetLabs/rpki-rs/pull/23